Update with ChainConfig

+++ core/src/chain_config.rs
@@ -4,6 +4,7 @@ use crate::crypto::Hasher;
 
 #[derive(Clone, Copy)]
 #[derive(Debug)]
+#[derive(PartialEq)]
 #[derive(Serialize, Deserialize)]
 pub struct ChainConfig {
     pub hasher: Hasher,
@@ -12,5 +13,5 @@ pub struct ChainConfig {
     pub confirmation_depth: u32,
     pub maintenance_window: u32,
     pub target_block_time_sec: u64,
-    pub max_vdf_difficulty_adjustment: u32,
+    pub max_vdf_difficulty_adjustment: f32,
 }
